[competitions] use new Pageable servise for competitions message info. Move some methods to view class

+++ src/main/java/com/dp_ua/iksparser/bot/command/impl/competition/CommandCompetitions.java
@@ -42,7 +42,7 @@ public class CommandCompetitions extends BaseCommand {
     @Override
     protected void perform(Message message) {
         String chatId = message.getChatId();
-        long commandArgument = getCommandArgument(message.getMessageText());
+        int commandArgument = getCommandArgument(message.getMessageText());
         try {
             competitionFacade.showCompetitions(chatId, commandArgument, message.getEditMessageId());
         } catch (ParsingException e) {
